feat: add logic of schema visibility
