+++ app/components/schema/schema.tsx
@@ -1,15 +1,72 @@
 import React, { useState, useRef, useCallback, useEffect } from 'react';
+import { useSelector } from 'react-redux';
+import {
+  buildClientSchema,
+  getIntrospectionQuery,
+  GraphQLSchema,
+  printSchema,
+} from 'graphql';
 
-import SchemaFetcher from './schemaSDL';
+import { RootState } from '../../slices/store';
 
 export default function SchemaPanel() {
   const [isOpen, setIsOpen] = useState<boolean>(false);
   const [panelWidth, setPanelWidth] = useState<number>(300);
   const panelRef = useRef<HTMLDivElement | null>(null);
+  const [sdlStatus, setSDLStatus] = useState<number>();
   const isResizing = useRef<boolean>(false);
+  const [schemaSDL, setSchemaSDL] = useState<string | null>(null);
+  const [error, setError] = useState<string | null>('');
+  const GRAPHQL_URL = useSelector((state: RootState) => state.schema.urlSdl);
+
+  useEffect(() => {
+    if (!GRAPHQL_URL.length) {
+      setIsOpen(false);
+    }
+  }, [GRAPHQL_URL]);
+
+  useEffect(() => {
+    const fetchSchema = async () => {
+      const introspectionQuery = getIntrospectionQuery();
+      setError('');
+      try {
+        const response = await fetch(GRAPHQL_URL, {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+          },
+          body: JSON.stringify({ query: introspectionQuery }),
+        });
+        setSDLStatus(response.status);
+        if (!response.ok) {
+          throw new Error(`HTTP error! status: ${response.status}`);
+        }
+        const result = await response.json();
+        if (result.errors) {
+          throw new Error(
+            `Errors returned from GraphQL: ${JSON.stringify(result.errors)}`
+          );
+        }
+        const schemaData = result.data;
+        if (!schemaData || !schemaData.__schema) {
+          throw new Error('Invalid introspection response: missing __schema');
+        }
+
+        const schema: GraphQLSchema = buildClientSchema(schemaData);
+        setSchemaSDL(printSchema(schema).replace(/"""/g, ''));
+      } catch (error: unknown) {
+        setError((error as Error).message);
+      }
+    };
+    fetchSchema();
+  }, [schemaSDL, GRAPHQL_URL, sdlStatus]);
 
   const togglePanel = () => {
-    setIsOpen((prev) => !prev);
+    if (sdlStatus === 200 && schemaSDL && GRAPHQL_URL.length) {
+      setIsOpen((prev) => !prev);
+    } else {
+      setIsOpen(false);
+    }
   };
 
   const handleMouseMove = useCallback((event: MouseEvent) => {
@@ -58,11 +115,16 @@ export default function SchemaPanel() {
             className="absolute left-0 top-0 bottom-0 w-2 cursor-ew-resize bg-gray-300"
             onMouseDown={startResize}
           />
-          <SchemaFetcher />
+          <h1>GraphQL Schema</h1>
+          {error ? (
+            <p className="text-red-500">{error}</p>
+          ) : (
+            <pre className=" h-full">{schemaSDL}</pre>
+          )}
         </div>
         <button
           onClick={togglePanel}
-          className="absolute py-1 px-2 z-50 left-[-70px] top-1/4 transform -translate-y-1/2 bg-[#fe6d12] text-white flex items-center justify-center shadow-md"
+          className={`absolute py-1 px-2 z-50 left-[-70px] top-1/4 transform -translate-y-1/2 bg-[${sdlStatus === 200 && schemaSDL && GRAPHQL_URL.length ? '#fe6d12' : '#fbc511'}] text-white flex items-center justify-center shadow-md`}
           style={{ transform: 'rotate(-90deg)' }}
         >
           Schema
