+++ app/slices/store.ts
@@ -2,11 +2,13 @@ import { configureStore } from '@reduxjs/toolkit';
 
 import headersReducer from './headersSlice';
 import variablesReducer from './variablesSlice';
+import schemaReducer from './sdlSlice';
 
 export const store = configureStore({
   reducer: {
     headers: headersReducer,
     variables: variablesReducer,
+    schema: schemaReducer,
   },
 });
